refactor: relocate repo and user modules
